Filter extracted paths to files

+++ src/datasets/utils/download_manager.py
@@ -294,7 +294,7 @@ class DownloadManager:
     def delete_extracted_files(self):
         paths_to_delete = set(self.extracted_paths.values()) - set(self.downloaded_paths.values())
         for key, path in list(self.extracted_paths.items()):
-            if path in paths_to_delete:
+            if path in paths_to_delete and os.path.isfile(path):
                 os.remove(path)
                 del self.extracted_paths[key]
